Update: 1.18
The "Export to PDF" button has been changed to a new name "Print Label".
Added the ability for the user to independently select the installed printer.
Now the saved PDF file after saving has a different name, in its name in the first part is the product number, in the second part is the label size.
Added application logo.